# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "User API",
        "version": "1.0.0",
        "description": "API for managing users"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": ["http", "https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
        "/user/create": {
            "post": {
                "summary": "Create a new user",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "firstName": { "type": "string" },
                                "lastName": { "type": "string" },
                                "email": { "type": "string" },
                                "password": { "type": "string" }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "summary": "User login",
                "parameters": [
                    {
                        "name": "credentials",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": { "type": "string" },
                                "password": { "type": "string" }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User logged in successfully"
                    },
                    "400": {
                        "description": "Invalid credentials"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    }
}